test: split isISO8601 validator tests
