+++ test/validators/isISO8601.test.js
@@ -1,0 +1,227 @@
+import test from '../testFunctions';
+
+const validISO8601 = [
+  '2009-12T12:34',
+  '2009',
+  '2009-05-19',
+  '2009-05-19',
+  '20090519',
+  '2009123',
+  '2009-05',
+  '2009-123',
+  '2009-222',
+  '2009-001',
+  '2009-W01-1',
+  '2009-W51-1',
+  '2009-W511',
+  '2009-W33',
+  '2009W511',
+  '2009-05-19',
+  '2009-05-19 00:00',
+  '2009-05-19 14',
+  '2009-05-19 14:31',
+  '2009-05-19 14:39:22',
+  '2009-05-19T14:39Z',
+  '2009-W21-2',
+  '2009-W21-2T01:22',
+  '2009-139',
+  '2009-05-19 14:39:22-06:00',
+  '2009-05-19 14:39:22+0600',
+  '2009-05-19 14:39:22-01',
+  '20090621T0545Z',
+  '2007-04-06T00:00',
+  '2007-04-05T24:00',
+  '2010-02-18T16:23:48.5',
+  '2010-02-18T16:23:48,444',
+  '2010-02-18T16:23:48,3-06:00',
+  '2010-02-18T16:23.4',
+  '2010-02-18T16:23,25',
+  '2010-02-18T16:23.33+0600',
+  '2010-02-18T16.23334444',
+  '2010-02-18T16,2283',
+  '2009-05-19 143922.500',
+  '2009-05-19 1439,55',
+  '2009-10-10',
+  '2020-366',
+  '2000-366',
+];
+
+const invalidISO8601 = [
+  '200905',
+  '2009367',
+  '2009-',
+  '2007-04-05T24:50',
+  '2009-000',
+  '2009-M511',
+  '2009M511',
+  '2009-05-19T14a39r',
+  '2009-05-19T14:3924',
+  '2009-0519',
+  '2009-05-1914:39',
+  '2009-05-19 14:',
+  '2009-05-19r14:39',
+  '2009-05-19 14a39a22',
+  '200912-01',
+  '2009-05-19 14:39:22+06a00',
+  '2009-05-19 146922.500',
+  '2010-02-18T16.5:23.35:48',
+  '2010-02-18T16:23.35:48',
+  '2010-02-18T16:23.35:48.45',
+  '2009-05-19 14.5.44',
+  '2010-02-18T16:23.33.600',
+  '2010-02-18T16,25:23:48,444',
+  '2010-13-1',
+  'nonsense2021-01-01T00:00:00Z',
+  '2021-01-01T00:00:00Znonsense',
+];
+
+describe('isISO8601', () => {
+  it('should validate ISO 8601 dates', () => {
+    // from http://www.pelagodesign.com/blog/2009/05/20/iso-8601-date-validation-that-doesnt-suck/
+    test({
+      validator: 'isISO8601',
+      valid: validISO8601,
+      invalid: invalidISO8601,
+    });
+  });
+
+  it('should validate ISO 8601 dates, with strict = true (regression)', () => {
+    test({
+      validator: 'isISO8601',
+      args: [
+        { strict: true },
+      ],
+      valid: validISO8601,
+      invalid: invalidISO8601,
+    });
+  });
+
+  it('should validate ISO 8601 dates, with strict = true', () => {
+    test({
+      validator: 'isISO8601',
+      args: [
+        { strict: true },
+      ],
+      valid: [
+        '2000-02-29',
+        '2009-123',
+        '2009-222',
+        '2020-366',
+        '2400-366',
+      ],
+      invalid: [
+        '2010-02-30',
+        '2009-02-29',
+        '2009-366',
+        '2019-02-31',
+      ],
+    });
+  });
+
+  it('should validate ISO 8601 dates, with strictSeparator = true', () => {
+    test({
+      validator: 'isISO8601',
+      args: [
+        { strictSeparator: true },
+      ],
+      valid: [
+        '2009-12T12:34',
+        '2009',
+        '2009-05-19',
+        '2009-05-19',
+        '20090519',
+        '2009123',
+        '2009-05',
+        '2009-123',
+        '2009-222',
+        '2009-001',
+        '2009-W01-1',
+        '2009-W51-1',
+        '2009-W511',
+        '2009-W33',
+        '2009W511',
+        '2009-05-19',
+        '2009-05-19T14:39Z',
+        '2009-W21-2',
+        '2009-W21-2T01:22',
+        '2009-139',
+        '20090621T0545Z',
+        '2007-04-06T00:00',
+        '2007-04-05T24:00',
+        '2010-02-18T16:23:48.5',
+        '2010-02-18T16:23:48,444',
+        '2010-02-18T16:23:48,3-06:00',
+        '2010-02-18T16:23.4',
+        '2010-02-18T16:23,25',
+        '2010-02-18T16:23.33+0600',
+        '2010-02-18T16.23334444',
+        '2010-02-18T16,2283',
+        '2009-10-10',
+        '2020-366',
+        '2000-366',
+      ],
+      invalid: [
+        '200905',
+        '2009367',
+        '2009-',
+        '2007-04-05T24:50',
+        '2009-000',
+        '2009-M511',
+        '2009M511',
+        '2009-05-19T14a39r',
+        '2009-05-19T14:3924',
+        '2009-0519',
+        '2009-05-1914:39',
+        '2009-05-19 14:',
+        '2009-05-19r14:39',
+        '2009-05-19 14a39a22',
+        '200912-01',
+        '2009-05-19 14:39:22+06a00',
+        '2009-05-19 146922.500',
+        '2010-02-18T16.5:23.35:48',
+        '2010-02-18T16:23.35:48',
+        '2010-02-18T16:23.35:48.45',
+        '2009-05-19 14.5.44',
+        '2010-02-18T16:23.33.600',
+        '2010-02-18T16,25:23:48,444',
+        '2010-13-1',
+        '2009-05-19 00:00',
+        // Previously valid cases
+        '2009-05-19 14',
+        '2009-05-19 14:31',
+        '2009-05-19 14:39:22',
+        '2009-05-19 14:39:22-06:00',
+        '2009-05-19 14:39:22+0600',
+        '2009-05-19 14:39:22-01',
+      ],
+    });
+  });
+
+  it('should validate ISO 8601 dates, with strict = true and strictSeparator = true (regression)', () => {
+    test({
+      validator: 'isISO8601',
+      args: [
+        { strict: true, strictSeparator: true },
+      ],
+      valid: [
+        '2000-02-29',
+        '2009-123',
+        '2009-222',
+        '2020-366',
+        '2400-366',
+      ],
+      invalid: [
+        '2010-02-30',
+        '2009-02-29',
+        '2009-366',
+        '2019-02-31',
+        '2009-05-19 14',
+        '2009-05-19 14:31',
+        '2009-05-19 14:39:22',
+        '2009-05-19 14:39:22-06:00',
+        '2009-05-19 14:39:22+0600',
+        '2009-05-19 14:39:22-01',
+      ],
+    });
+  });
+});
